worklog-backend/requirements.txt
Django~=4.2.0
djangorestframework~=3.14.0
Pillow~=10.0.0             # For image processing in Django models (often a dependency)
python-dotenv~=1.0.0      # For loading environment variables from a .env file
dj-database-url~=1.2.0    # For parsing database URLs (useful for deployment)
psycopg2-binary~=2.9.9    # PostgreSQL adapter (install if using PostgreSQL)
mysqlclient~=2.2.0        # MySQL adapter (install if using MySQL)
gunicorn~=21.2.0         # Production WSGI HTTP server
whitenoise~=6.6.0         # For serving static files in production
django-cors-headers~=4.3.0 # For Cross-Origin Resource Sharing (CORS)
djangorestframework-simplejwt~=5.3.0 # For JSON Web Token (JWT) authentication
